requirements-docs.txt
sphinx<2.2
sphinx-click